requirements.txt
aiofiles==23.2.1
aerich==0.6.1
Pillow==8.4.0
aiohttp==3.7.4.post0
aioredis==2.0.0a1
aliyun-log-python-sdk==0.7.3
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.15.0
APScheduler==3.7.0
asgiref==3.4.1
async-timeout==3.0.1
asyncoss==0.0.4
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
crcmod==1.7
cryptography==3.4.7
dateparser==1.1.0
Deprecated==1.2.13
Django==4.0
elasticsearch==7.16.1
fastapi==0.67.0
h11==0.12.0
hiredis==2.0.0
idna==3.2
image==1.5.33
jmespath==0.10.0
multidict==5.1.0
oss2==2.15.0
protobuf==3.19.1
pycparser==2.20
pydantic==1.8.2
pyftpdlib==1.5.6
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.1
qrcode==7.3.1
redis==4.0.2
regex==2021.11.10
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
starlette==0.14.2
typing-extensions
tzlocal==2.1
urllib3==1.26.6
uvicorn==0.14.0
wrapt==1.13.3
yarl==1.6.3
pdfkit==1.0.0
aiomysql==0.0.22
tortoise-orm
python-jose==3.3.0
passlib==1.7.4
pycryptodome==3.20.0

